floating-point

+++ src/main/java/com/Nxer/TwistSpaceTechnology/common/machine/multiMachineClasses/GTCM_MultiMachineBase.java
@@ -140,7 +140,7 @@ public abstract class GTCM_MultiMachineBase<T extends GTCM_MultiMachineBase<T>>
      */
     @Override
     protected long getActualEnergyUsage() {
-        return -this.lEUt * (10000 / Math.max(1000, mEfficiency));
+        return (long) (-this.lEUt * (10000.0 / Math.max(1000, mEfficiency)));
     }
 
     /**
